+++ carp_mobile_sensing/lib/runtime/notification/awesome_notification_controller.dart
@@ -124,6 +124,7 @@ class AwesomeNotificationController implements NotificationController {
       UserTask? task = AppTaskController().getUserTask(payload);
       info('NotificationController - User Task notification selected - $task');
       if (task != null) {
+        task.state = UserTaskState.notified;
         task.onNotification();
       } else {
         warning(