+++ packages/carp_survey_package/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 
 * update to `carp_mobile_sensing` v. 1.1.0
+* update to the new `UserTask` model where the app (and not this package) is responsible for showing the user interface (`Widget`) of the survey task.
 
 ## 0.40.1